# acme-hotel-api/docs/openapi.yaml
openapi: 3.1.0
info:
  title: ACME Hotel API
  description: A comprehensive hotel booking API providing hotel search, detailed
    information, and full booking management. Designed to be consumed by developers
    and MCP tool servers.
  version: 1.0.0
paths:
...
  /hotels:
    get:
      tags:
      - Hotels
      summary: Search hotels
      description: Search and filter hotels across all cities. Returns hotel summaries
        (without full reviews). All filters are optional and combine with AND logic.
        Use `q` for free-text search across name, city, country, and description.
      operationId: searchHotels
      parameters:
      - name: q
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Free-text search across name, city, country, and description
          title: Q
        description: Free-text search across name, city, country, and description
      - name: city
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by city name (case-insensitive)
          examples:
          - Paris
          title: City
        description: Filter by city name (case-insensitive)
      - name: country
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by country (case-insensitive)
          examples:
          - France
          title: Country
        description: Filter by country (case-insensitive)
      - name: min_price
        in: query
        required: false
        schema:
          anyOf:
          - type: number
            minimum: 0
          - type: 'null'
          description: Minimum nightly price (USD) across any room type
          title: Min Price
        description: Minimum nightly price (USD) across any room type
      - name: max_price
        in: query
        required: false
        schema:
          anyOf:
          - type: number
            minimum: 0
          - type: 'null'
          description: Maximum nightly price (USD) across any room type
          title: Max Price
        description: Maximum nightly price (USD) across any room type
      - name: min_rating
        in: query
        required: false
        schema:
          anyOf:
          - type: number
            maximum: 5
            minimum: 0
          - type: 'null'
          description: "Minimum guest rating (0\u20135)"
          title: Min Rating
        description: "Minimum guest rating (0\u20135)"
      - name: stars
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
            maximum: 5
            minimum: 1
          - type: 'null'
          description: "Exact star rating (1\u20135)"
          title: Stars
        description: "Exact star rating (1\u20135)"
      - name: amenity
        in: query
        required: false
        schema:
          anyOf:
          - type: array
            items:
              type: string
          - type: 'null'
          description: Required amenities (all must be present)
          examples:
          - - spa
            - swimming_pool
          title: Amenity
        description: Required amenities (all must be present)
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HotelSummary'
                title: Response Searchhotels
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /hotels/{hotel_id}:
    get:
      tags:
      - Hotels
      summary: Get hotel details
      description: Returns full hotel details including room types, reviews, and contact
        information.
      operationId: getHotelById
      parameters:
      - name: hotel_id
        in: path
        required: true
        schema:
          type: string
          title: Hotel Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Hotel'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /hotels/{hotel_id}/reviews:
    get:
      tags:
      - Hotels
      summary: Get hotel reviews
      description: Returns all guest reviews for a specific hotel.
      operationId: getHotelReviews
      parameters:
      - name: hotel_id
        in: path
        required: true
        schema:
          type: string
          title: Hotel Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Review'
                title: Response Gethotelreviews
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /bookings:
    get:
      tags:
      - Bookings
      summary: List bookings
      description: Returns bookings for the authenticated user, identified by the
        X-User-Email header.
      operationId: listBookings
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Booking'
                type: array
                title: Response Listbookings
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Bookings
      summary: Create a booking
      description: Create a new hotel room booking. The total price is automatically
        calculated from the room type nightly rate and stay duration. Validates hotel
        existence, room type, dates, and guest capacity.
      operationId: createBooking
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /bookings/{booking_id}:
    get:
      tags:
      - Bookings
      summary: Get booking details
      description: Returns full details of a specific booking by its reference ID.
      operationId: getBookingById
      parameters:
      - name: booking_id
        in: path
        required: true
        schema:
          type: string
          title: Booking Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - Bookings
      summary: Update a booking
      description: Modify an existing booking. Only confirmed bookings can be updated.
        Send only the fields you want to change. The total price is automatically
        recalculated if dates or room type change.
      operationId: updateBooking
      parameters:
      - name: booking_id
        in: path
        required: true
        schema:
          type: string
          title: Booking Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Bookings
      summary: Cancel a booking
      description: Cancel an existing booking. Only confirmed bookings can be cancelled.
        Returns the updated booking with status 'cancelled'.
      operationId: cancelBooking
      parameters:
      - name: booking_id
        in: path
        required: true
        schema:
          type: string
          title: Booking Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
...
    BookingUpdate:
      properties:
        room_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Room Type
          description: New room type
          examples:
          - suite
        check_in:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Check In
          description: New check-in date
          examples:
          - '2025-06-16'
        check_out:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Check Out
          description: New check-out date
          examples:
          - '2025-06-23'
        guests:
          anyOf:
          - type: integer
            minimum: 1.0
          - type: 'null'
          title: Guests
          description: Updated guest count
          examples:
          - 3
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
          description: Updated notes
      type: object
      title: BookingUpdate
      description: Request body to modify an existing booking. All fields optional.
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    Hotel:
      properties:
        id:
          type: string
          title: Id
          description: Unique hotel identifier
          examples:
          - grand-london
        name:
          type: string
          title: Name
          description: Hotel name
          examples:
          - The Grand London
        city:
          type: string
          title: City
          description: City
          examples:
          - London
        country:
          type: string
          title: Country
          description: Country
          examples:
          - United Kingdom
        address:
          type: string
          title: Address
          description: Full street address
        stars:
          type: integer
          maximum: 5.0
          minimum: 1.0
          title: Stars
          description: "Star rating (1\u20135)"
          examples:
          - 5
        description:
          type: string
          title: Description
          description: Detailed hotel description
        check_in_time:
          type: string
          title: Check In Time
          description: Check-in time (HH:MM)
          examples:
          - '15:00'
        check_out_time:
          type: string
          title: Check Out Time
          description: Check-out time (HH:MM)
          examples:
          - '11:00'
        amenities:
          items:
            type: string
          type: array
          title: Amenities
          description: List of amenities
          examples:
          - - free_wifi
            - spa
            - restaurant
        room_types:
          items:
            $ref: '#/components/schemas/RoomType'
          type: array
          title: Room Types
          description: Available room categories
        rating:
          type: number
          maximum: 5.0
          minimum: 0.0
          title: Rating
          description: Average guest rating
          examples:
          - 4.8
        review_count:
          type: integer
          title: Review Count
          description: Total number of reviews
          examples:
          - 2847
        reviews:
          items:
            $ref: '#/components/schemas/Review'
          type: array
          title: Reviews
          description: Selected guest reviews
        contact:
          $ref: '#/components/schemas/HotelContact'
      type: object
      required:
      - id
      - name
      - city
      - country
      - address
      - stars
      - description
      - check_in_time
      - check_out_time
      - amenities
      - room_types
      - rating
      - review_count
      - reviews
      - contact
      title: Hotel
      description: Full hotel details including rooms, reviews and contact info.
    HotelContact:
      properties:
        phone:
          type: string
          title: Phone
          description: Phone number
          examples:
          - +44 20 7123 4567
        email:
          type: string
          title: Email
          description: Reservation email
          examples:
          - reservations@grandlondon.com
      type: object
      required:
      - phone
      - email
      title: HotelContact
    HotelSummary:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        city:
          type: string
          title: City
        country:
          type: string
          title: Country
        address:
          type: string
          title: Address
        stars:
          type: integer
          title: Stars
        description:
          type: string
          title: Description
        check_in_time:
          type: string
          title: Check In Time
        check_out_time:
          type: string
          title: Check Out Time
        amenities:
          items:
            type: string
          type: array
          title: Amenities
        room_types:
          items:
            $ref: '#/components/schemas/RoomType'
          type: array
          title: Room Types
        rating:
          type: number
          title: Rating
        review_count:
          type: integer
          title: Review Count
        contact:
          $ref: '#/components/schemas/HotelContact'
      type: object
      required:
      - id
      - name
      - city
      - country
      - address
      - stars
      - description
      - check_in_time
      - check_out_time
      - amenities
      - room_types
      - rating
      - review_count
      - contact
      title: HotelSummary
      description: Lightweight hotel listing without reviews.
    Review:
      properties:
        author:
          type: string
          title: Author
          description: Reviewer name
          examples:
          - Sarah M.
        rating:
          type: integer
          maximum: 5.0
          minimum: 1.0
          title: Rating
          description: Rating from 1 to 5
          examples:
          - 5
        title:
          type: string
          title: Title
          description: Review title
          examples:
          - Absolutely stunning
        comment:
          type: string
          title: Comment
          description: Review text
        date:
          type: string
          title: Date
          description: Review date (YYYY-MM-DD)
          examples:
          - '2024-11-15'
      type: object
      required:
      - author
      - rating
      - title
      - comment
      - date
      title: Review
...